Fix log check after each
