+++ systemtest/src/test/java/io/strimzi/systemtest/ConnectST.java
@@ -9,7 +9,6 @@ import io.fabric8.kubernetes.api.model.Quantity;
 import io.fabric8.kubernetes.api.model.ResourceRequirementsBuilder;
 import io.strimzi.api.kafka.Crds;
 import io.strimzi.api.kafka.model.KafkaConnectResources;
-import io.strimzi.api.kafka.model.KafkaConnectorBuilder;
 import io.strimzi.api.kafka.model.KafkaResources;
 import io.strimzi.systemtest.clients.lib.KafkaClient;
 import io.strimzi.systemtest.utils.StUtils;
